fix file type association on Windows 7
The previous code probably assumed that Versions following Vista
would work with Vista’s API. Unfortunately, Windows 7 does not.
Using the old (NT <= 5.1) way appears to work as intended, so
use that for everything but Vista.

Perhaps the Vista API usage is just a gimmick that could be removed
(dealing with Windows API is Not Fun). I have no idea, and I simply
do not care enough about Vista to test that, but I don’t want to
break it either, so I keep it in there at least for now.