Workaround special case in Matroska chapter handling; Always assume UTF-8 output from mplayer2
When the first segment is an external segment, ID_CHAPTERS is not
necessarily equal to the count of chapters in the main segment.

Since my mplayer2 builds now set the console output codepage to UTF-8,
parsing chapter names and other things with characters outside the local
codepage can now work as expected.